LayoutTests/storage/domstorage/sessionstorage/no-quota-expected.txt

Test whether we can add >5mb (the default quota) of data to DOM Storage

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing window.sessionStorage
storage.clear()
PASS storage.length is 0
Creating 'data' which contains 64K of data
PASS data.length is 65536
Putting 'data' into 40 window.sessionStorage buckets.
Putting 'data' into another bucket.h
PASS Insertion worked.
PASS successfullyParsed is true

TEST COMPLETE
